requirements.txt
Django>=1.8,<=2.0.11
psycopg2-binary
pillow
requests
whitenoise==3.3.1
python-decouple
djangorestframework
markdown
django-filter
gunicorn
django-webpack-loader
dj-database-url
waitress
hellowebapp-deploy
djangorestframework-jwt==1.11.0
coreapi==2.3.3
pyyaml==5.1.1
django-cors-headers==2.2.0
django-rest-swagger==2.2.0
openapi-codec==1.3.2
simplejson==3.16.0